+++ babel.config.js
@@ -2,7 +2,15 @@ module.exports = function (api) {
   api.cache(true);
 
   const presets = [
-		"@babel/preset-env",
+		["@babel/preset-env", {
+			targets: {
+				edge: "17",
+				ie: '11',
+				firefox: "60",
+				chrome: "67",
+				safari: "11.1",
+			},
+    }],
 		"@babel/preset-react",
 	];