+++ src/choices-guide/index.jsx
@@ -1,8 +1,3 @@
-// polyfills
-import 'whatwg-fetch';
-import 'core-js/es/array/find';
-import 'custom-event-polyfill';
-
 // react
 import React from 'react';
 import ReactDOM from 'react-dom';