+++ src/modal-popup/index.jsx
@@ -1,7 +1,3 @@
-// polyfills
-import 'whatwg-fetch';
-import 'core-js/es/array/find-index';
-
 // react
 import React from 'react';
 import ReactDOM from 'react-dom';