+++ src/previous-next-button-block/index.jsx
@@ -1,6 +1,3 @@
-// polyfills
-import 'whatwg-fetch';
-
 // react
 import React from 'react';
 import ReactDOM from 'react-dom';